+++ .flake8
@@ -4,6 +4,10 @@ ignore = E128,E203,E501,W503
 per-file-ignores =
 	# module level import not at top of file
 	pumpkin.py:E402
+	# module/object imported but unused
+	pie/__init__.py:F401
+	pie/check.py:F401
+	pie/utils/__init__.py:F401
 exclude = .git,.venv
 count = True
 max-complexity = 25
