--- core/lang/cs.ini
@@ -1,13 +1,0 @@
-[help]
-no category = Nápověda
-module = Modul
-commands = Příkazy
-command not found = Žádný příkaz **((name))** neznám
-no subcommand = Příkaz **((name))** nemá žádné podpříkazy
-no named subcommand = Příkaz **((name))** nemá podpříkaz **((subcommand))**
-
-[create_embed]
-footer = Odpověď pro
-
-[version]
-reply = Příkaz není v této verzi dostupný
