--- core/lang/en.ini
@@ -1,13 +1,0 @@
-[help]
-no category = Help
-module = Module
-commands = Commands
-command not found = I don't know the **((name))** command
-no subcommand = Command **((name))** has no subcommand
-no named subcommand = Command **((name))** does not have subcommand **((subcommand))**
-
-[create_embed]
-footer = Answer for
-
-[version]
-reply = This command is not available in this version
