+++ pie/check.py
@@ -1,0 +1,2 @@
+from pie.acl import acl
+from pie.spamchannel import spamchannel
