+++ pie/logger/database.py
@@ -3,7 +3,7 @@ from typing import Optional, List, Dict
 
 from sqlalchemy import BigInteger, Column, String, Integer
 
-from database import database, session
+from pie.database import database, session
 
 
 class LogConf(database.base):
